requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
clstm==0.0.5
future==0.16.0
idna==2.7
img2pdf==0.3.1
Jinja2==2.10
kraken==0.9.16
lxml==4.2.4
MarkupSafe==1.0
numpy==1.15.1
Pillow==5.2.0
protobuf==3.6.1
pytesseract==0.2.4
python-bidi==0.4.0
regex==2018.8.29
requests==2.19.1
scipy==1.1.0
six==1.11.0
urllib3==1.23
